myBatis 관련 파일 삭제
