+++ build.gradle
@@ -21,8 +21,7 @@ repositories {
 dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-web'
 	implementation 'org.springframework.boot:spring-boot-starter-validation'
-	implementation 'org.mybatis.spring.boot:mybatis-spring-boot-starter' // MyBatis
-	implementation 'org.springframework.boot:spring-boot-starter-data-jpa' // JPA
+	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	implementation 'org.bgee.log4jdbc-log4j2:log4jdbc-log4j2-jdbc4.1:1.16'
 	implementation 'com.github.pagehelper:pagehelper-spring-boot-starter:1.4.6' // 페이징(like Pageable of JPA)
 	implementation 'org.apache.commons:commons-lang3:3.12.0'
