+++ src/main/java/flab/buynow/config/DataSourceConfig.java
@@ -1,0 +1,26 @@
+package flab.buynow.config;
+
+import com.zaxxer.hikari.HikariConfig;
+import com.zaxxer.hikari.HikariDataSource;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import javax.sql.DataSource;
+
+@Configuration
+public class DataSourceConfig {
+
+    @Bean
+    @ConfigurationProperties(prefix = "spring.datasource.hikari")
+    public HikariConfig hikariConfig() {
+        return new HikariConfig();
+    }
+
+    @Bean
+    public DataSource dataSource() {
+        DataSource dataSource = new HikariDataSource(hikariConfig());
+        return dataSource;
+    }
+
+}
