--- src/main/java/flab/buynow/config/MyBatisConfig.java
@@ -1,56 +1,0 @@
-package flab.buynow.config;
-
-import com.zaxxer.hikari.HikariConfig;
-import com.zaxxer.hikari.HikariDataSource;
-import org.apache.ibatis.session.SqlSessionFactory;
-import org.mybatis.spring.SqlSessionFactoryBean;
-import org.mybatis.spring.SqlSessionTemplate;
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.io.support.PathMatchingResourcePatternResolver;
-
-import javax.sql.DataSource;
-import java.io.IOException;
-
-@Configuration
-@MapperScan(basePackages = "flab.buynow")
-public class MyBatisConfig {
-
-    private final ApplicationContext applicationContext;
-
-    public MyBatisConfig(ApplicationContext applicationContext) {
-        this.applicationContext = applicationContext;
-    }
-
-    @Bean
-    @ConfigurationProperties(prefix = "spring.datasource.hikari")
-    public HikariConfig hikariConfig() {
-        return new HikariConfig();
-    }
-
-    @Bean
-    public DataSource dataSource() {
-        DataSource dataSource = new HikariDataSource(hikariConfig());
-        return dataSource;
-    }
-
-    @Bean
-    public SqlSessionFactoryBean sqlSessionFactory(DataSource dataSource) throws IOException {
-        SqlSessionFactoryBean factoryBean = new SqlSessionFactoryBean();
-        factoryBean.setDataSource(dataSource);
-        factoryBean.setMapperLocations(
-            new PathMatchingResourcePatternResolver().getResources("classpath:/mapper/*.xml"));
-        factoryBean.setConfigLocation(
-            applicationContext.getResource("classpath:mybatis-config.xml"));
-        return factoryBean;
-    }
-
-    @Bean
-    public SqlSessionTemplate sqlSession(SqlSessionFactory sqlSessionFactory) {
-        return new SqlSessionTemplate(sqlSessionFactory);
-    }
-
-}
