--- src/main/resources/mapper/itemMapper.xml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="flab.buynow.item.repository.ItemMyBatisRepository">
-
-    <select id="findById" parameterType="java.lang.Long" resultType="flab.buynow.item.domain.Item">
-        SELECT
-            id
-            ,name
-            ,price
-            ,sale_price
-            ,stock
-            ,start_date
-            ,end_date
-        FROM item
-        WHERE id = #{id}
-        FOR UPDATE
-    </select>
-
-    <select id="findAll" parameterType="flab.buynow.common.dto.PageInfoDto" resultType="flab.buynow.item.domain.Item">
-        SELECT
-            id
-            ,name
-            ,price
-            ,sale_price
-            ,stock
-            ,start_date
-            ,end_date
-        FROM item
-        WHERE id > #{offset}
-        ORDER BY id ASC
-        LIMIT #{limit}
-    </select>
-
-    <insert id="create" parameterType="flab.buynow.item.domain.Item">
-        INSERT INTO item(
-            name
-            ,price
-            ,sale_price
-            ,stock
-            ,start_date
-            ,end_date
-        ) VALUES(
-            #{name}
-            ,#{price}
-            ,#{salePrice}
-            ,#{stock}
-            ,#{startDate}
-            ,#{endDate}
-        )
-    </insert>
-
-    <update id="update" parameterType="flab.buynow.item.domain.Item">
-        UPDATE item SET
-            name = #{name}
-            ,price = #{price}
-            ,sale_price = #{salePrice}
-            ,stock = #{stock}
-            ,start_date = #{startDate}
-            ,end_date = #{endDate}
-        WHERE id = #{id}
-    </update>
-
-    <update id="minusStock" parameterType="flab.buynow.item.domain.Item">
-        UPDATE item SET
-            stock = stock - #{stock}
-        WHERE id = #{id}
-    </update>
-
-    <update id="plusStock" parameterType="flab.buynow.item.domain.Item">
-        UPDATE item SET
-            stock = stock + #{stock}
-        WHERE id = #{id}
-    </update>
-
-</mapper>
