--- src/main/resources/mapper/memberMapper.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="flab.buynow.member.repository.MemberMyBatisRepository">
-
-    <select id="findByLoginId" parameterType="java.lang.String" resultType="flab.buynow.member.domain.Member">
-        SELECT
-            id
-            ,login_id
-            ,password
-            ,name
-            ,telephone_number
-            ,address
-            ,address_detail
-            ,join_date
-            ,last_updated_date
-            ,admin_yn
-        FROM member
-        WHERE login_id = #{loginId}
-    </select>
-
-    <select id="findAll" parameterType="flab.buynow.member.dto.PageInfoDto" resultType="flab.buynow.member.domain.Member">
-        SELECT
-            id
-            ,login_id
-            ,password
-            ,name
-            ,telephone_number
-            ,address
-            ,address_detail
-            ,join_date
-            ,last_updated_date
-        FROM member
-        WHERE id > #{offset}
-        ORDER BY id ASC
-        LIMIT #{limit}
-    </select>
-
-    <insert id="create" parameterType="flab.buynow.member.domain.Member">
-        INSERT INTO member(
-            login_id
-            ,password
-            ,name
-            ,telephone_number
-            ,address
-            ,address_detail
-            ,join_date
-        ) VALUES(
-            #{loginId}
-            ,#{password}
-            ,#{name}
-            ,#{telephoneNumber}
-            ,#{address}
-            ,#{addressDetail}
-            ,NOW()
-        )
-    </insert>
-
-    <update id="update" parameterType="flab.buynow.member.domain.Member">
-        UPDATE member SET
-            password = #{password}
-            ,name = #{name}
-            ,telephone_number = #{telephoneNumber}
-            ,address = #{address}
-            ,address_detail = #{addressDetail}
-            ,last_updated_date = NOW()
-        WHERE login_id = #{loginId}
-    </update>
-
-</mapper>
