--- src/main/resources/mapper/orderMapper.xml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="flab.buynow.order.repository.OrderRepository">
-
-    <select id="findById" parameterType="java.lang.Long" resultType="flab.buynow.order.domain.Orders">
-        SELECT
-            id,
-            user_id,
-            order_no,
-            name,
-            telephone_number,
-            address,
-            address_detail,
-            order_date,
-            status
-        FROM orders
-        WHERE id = #{id}
-    </select>
-
-    <select id="findAll" parameterType="flab.buynow.common.dto.PageInfoDto" resultType="flab.buynow.order.domain.Orders">
-        SELECT
-            id,
-            user_id,
-            order_no,
-            name,
-            telephone_number,
-            address,
-            address_detail,
-            order_date,
-            status
-        FROM orders
-        WHERE id > #{offset}
-        ORDER BY id ASC
-            LIMIT #{limit}
-    </select>
-
-    <insert id="create" parameterType="flab.buynow.order.domain.Orders" useGeneratedKeys="true" keyProperty="id">
-        INSERT INTO orders(
-            user_id
-            ,order_no
-            ,name
-            ,telephone_number
-            ,address
-            ,address_detail
-            ,order_date
-            ,status
-        ) VALUES(
-            #{userId}
-            ,#{orderNo}
-            ,#{name}
-            ,#{telephoneNumber}
-            ,#{address}
-            ,#{addressDetail}
-            ,NOW()
-            ,#{status}
-        )
-    </insert>
-
-    <update id="update" parameterType="flab.buynow.order.domain.Orders">
-        UPDATE orders SET
-            name = #{name}
-            ,telephone_number = #{telephoneNumber}
-            ,address = #{address}
-            ,address_detail = #{addressDetail}
-        WHERE id = #{id}
-    </update>
-
-    <update id="cancel" parameterType="java.lang.Long">
-        UPDATE orders SET
-            status = 'CANCEL'
-            ,cancel_date = NOW()
-        WHERE id = #{id}
-    </update>
-
-</mapper>
